[MOJO-2066] Use Java 6.
It needs some extra work to get the plugin compiled with Java 1.5. I
would like to avoid this by setting the JVM requirement to Java 6.
(Maven itself runs with Java 1.6 since MNG-5558.) This should be done
before the 1.0.0 release is created, because this change should not be
made in the 1.x.x line for compatibility reasons.

git-svn-id: file:///home/tiste/MOJOHAUS-TO-GIT/SVN-MOJO-WIP/trunk/mojo/tidy-maven-plugin@20386 52ab4f32-60fc-0310-b215-8acea882cd1b