Fix Travis setup
"openjdk7" and "oraclejdk8" are not supported without further modifications on Travis anymore,
therefore replacing them with OpenJDK 8 and 11.

Removing tests for Maven 3.0.5 as it connects to Maven Central via HTTP which is not supported
anymore as of 2020-01-15, see https://links.sonatype.com/central/501-https-required.

Additionally using the latest Maven version and upgrading the Maven wrapper.